Formatting.Holey

Description

Copy of the holey monoids library but with constructor exported.

Formatting library

type Format r x = Holey Builder r x

A formatter.

type Holey m r a = HoleyT r a m

newtype HoleyT r a m

The type of a monoid with holes. The underlying monoid is represented by type parameter m. The r is the result type and stays polymorphic until the very last moment when run is called. The last argument a is always a function with zero or more arguments, finally resulting in r. Ordering the arguments in this order allows holey monoids to be composed using ., stacking the expected arguments. Note that the Monoid constraint is only used in the identity Holey and in composing two Holeys.

Constructors

Holey

Fields

runHM :: (m -> r) -> a

(%) :: Format b c -> Format r b -> Format r c infixr 9

Composition operator. The same as category composition.

(%.) :: Format r (Builder -> b) -> Format b c -> Format r c infixr 8

Function compose two holeys. Will feed the result of one holey into another.

now :: Builder -> Format r r

Insert a constant monoidal value.

bind :: Format b c -> (Builder -> Format a b) -> Format a c

Monadic indexed bind for holey monoids.

later :: (a -> Builder) -> Format r (a -> r)

Insert a monoidal value that is not specified until the computation is run. The argument that is expected later is converted to the monoid type using the given conversion function.