Formatting.Internal

Description

Internal format starters.

Documentation

format :: Format Text a -> a

Run the formatter and return a lazy Text value.

sformat :: Format Text a -> a

Run the formatter and return a strict Text value.

bprint :: Format Builder a -> a

Run the formatter and return a Builder value.

fprint :: Format (IO ()) a -> a

Run the formatter and print out the text to stdout.

hprint :: Handle -> Format (IO ()) a -> a

Run the formatter and put the output onto the given Handle.

formatToString :: Format [Char] a -> a

Run the formatter and return a list of characters.