Formatting.Time

Description

Formatters for time.

For TimeZone (and ZonedTime and UTCTime):

tz :: FormatTime a => Format r (a -> r)

Timezone offset on the format -HHMM.

tzName :: FormatTime a => Format r (a -> r)

Timezone name.

datetime :: FormatTime a => Format r (a -> r)

As dateTimeFmt locale (e.g. %a %b %e %H:%M:%S %Z %Y).

For TimeOfDay (and LocalTime and ZonedTime and UTCTime):

hm :: FormatTime a => Format r (a -> r)

Same as %H:%M.

hms :: FormatTime a => Format r (a -> r)

Same as %H:%M:%S.

hmsL :: FormatTime a => Format r (a -> r)

As timeFmt locale (e.g. %H:%M:%S).

hmsPL :: FormatTime a => Format r (a -> r)

As time12Fmt locale (e.g. %I:%M:%S %p).

dayHalf :: FormatTime a => Format r (a -> r)

Day half from (amPm locale), converted to lowercase, am, pm.

dayHalfU :: FormatTime a => Format r (a -> r)

Day half from (amPm locale), AM, PM.

hour24 :: FormatTime a => Format r (a -> r)

Hour, 24-hour, leading 0 as needed, 00 - 23.

hour12 :: FormatTime a => Format r (a -> r)

Hour, 12-hour, leading 0 as needed, 01 - 12.

hour24S :: FormatTime a => Format r (a -> r)

Hour, 24-hour, leading space as needed, 0 - 23.

hour12S :: FormatTime a => Format r (a -> r)

Hour, 12-hour, leading space as needed, 1 - 12.

minute :: FormatTime a => Format r (a -> r)

Minute, 00 - 59.

second :: FormatTime a => Format r (a -> r)

Second, without decimal part, 00 - 60.

pico :: FormatTime a => Format r (a -> r)

Picosecond, including trailing zeros, 000000000000 - 999999999999.

decimals :: FormatTime a => Format r (a -> r)

Decimal point and up to 12 second decimals, without trailing zeros. For a whole number of seconds, this produces the empty string.

For UTCTime and ZonedTime

epoch :: FormatTime a => Format r (a -> r)

For Day (and LocalTime and ZonedTime and UTCTime):

dateSlash :: FormatTime a => Format r (a -> r)

Same as %m/%d/%y.

dateDash :: FormatTime a => Format r (a -> r)

Same as %Y-%m-%d.

dateSlashL :: FormatTime a => Format r (a -> r)

As dateFmt locale (e.g. %m/%d/%y).

year :: FormatTime a => Format r (a -> r)

Year.

yy :: FormatTime a => Format r (a -> r)

Last two digits of year, 00 - 99.

century :: FormatTime a => Format r (a -> r)

Century (being the first two digits of the year), 00 - 99.

monthName :: FormatTime a => Format r (a -> r)

Month name, long form (fst from months locale), January - December.

monthNameShort :: FormatTime a => Format r (a -> r)

%H] month name, short form (snd from months locale), Jan - Dec@.

month :: FormatTime a => Format r (a -> r)

Month of year, leading 0 as needed, 01 - 12.

dayOfMonth :: FormatTime a => Format r (a -> r)

Day of month, leading 0 as needed, 01 - 31.

dayOfMonthOrd :: FormatTime a => Format r (a -> r)

Day of month, 1st, 2nd, 25th, etc.

dayOfMonthS :: FormatTime a => Format r (a -> r)

Day of month, leading space as needed, 1 - 31.

day :: FormatTime a => Format r (a -> r)

Day of year for Ordinal Date format, 001 - 366.

weekYear :: FormatTime a => Format r (a -> r)

Year for Week Date format e.g. 2013.

weekYY :: FormatTime a => Format r (a -> r)

Last two digits of year for Week Date format, 00 - 99.

weekCentury :: FormatTime a => Format r (a -> r)

Century (first two digits of year) for Week Date format, 00 - 99.

week :: FormatTime a => Format r (a -> r)

Week for Week Date format, 01 - 53.

dayOfWeek :: FormatTime a => Format r (a -> r)

Day for Week Date format, 1 - 7.

dayNameShort :: FormatTime a => Format r (a -> r)

Day of week, short form (snd from wDays locale), Sun - Sat.

dayName :: FormatTime a => Format r (a -> r)

Day of week, long form (fst from wDays locale), Sunday - Saturday.

weekFromZero :: FormatTime a => Format r (a -> r)

Week number of year, where weeks start on Sunday (as sundayStartWeek), 00 - 53.

dayOfWeekFromZero :: FormatTime a => Format r (a -> r)

Day of week number, 0 (= Sunday) - 6 (= Saturday).

weekOfYearMon :: FormatTime a => Format r (a -> r)

Week number of year, where weeks start on Monday (as mondayStartWeek), 00 - 53.

Time spans, diffs, NominalDiffTime, DiffTime, etc.

diff

Arguments

:: RealFrac n

=> Bool
Display 'in/ago'?

-> Format r (n -> r)
Example: '3 seconds ago', 'in three days'.)

Display a time span as one time relative to another. Input is assumed to be seconds. Typical inputs are NominalDiffTime and DiffTime.

years

Arguments

:: RealFrac n

=> Int
Decimal places.

-> Format r (n -> r)

Display the absolute value time span in years.

days

Arguments

:: RealFrac n

=> Int
Decimal places.

-> Format r (n -> r)

Display the absolute value time span in days.

hours

Arguments

:: RealFrac n

=> Int
Decimal places.

-> Format r (n -> r)

Display the absolute value time span in hours.

minutes

Arguments

:: RealFrac n

=> Int
Decimal places.

-> Format r (n -> r)

Display the absolute value time span in minutes.

seconds

Arguments

:: RealFrac n

=> Int
Decimal places.

-> Format r (n -> r)

Display the absolute value time span in seconds.

Internal.

fmt :: FormatTime a => Text -> a -> Text

Formatter call. Probably don't want to use this.